+++ backend/LexBoxApi/Controllers/AdminController.cs
@@ -1,0 +1,42 @@
+using LexBoxApi.Auth;
+using LexCore;
+using LexCore.Auth;
+using LexData;
+using Microsoft.AspNetCore.Authentication;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+
+namespace LexBoxApi.Controllers;
+
+[ApiController]
+[Route("/api/login")]
+public class AdminController : ControllerBase
+{
+    private readonly LexAuthService _lexAuthService;
+    private readonly LexBoxDbContext _lexBoxDbContext;
+    private readonly LoggedInContext _loggedInContext;
+
+    public AdminController(LexAuthService lexAuthService,
+        LexBoxDbContext lexBoxDbContext,
+        LoggedInContext loggedInContext)
+    {
+        _lexAuthService = lexAuthService;
+        _lexBoxDbContext = lexBoxDbContext;
+        _loggedInContext = loggedInContext;
+    }
+
+    public record ResetPasswordAdminRequest(string PasswordHash, Guid userId);
+
+    [HttpPost("resetPasswordAdmin")]
+    [AdminRequired]
+    public async Task<ActionResult> ResetPasswordAdmin(ResetPasswordAdminRequest request)
+    {
+        var passwordHash = request.PasswordHash;
+        var lexAuthUser = _loggedInContext.User;
+        var user = await _lexBoxDbContext.Users.FirstAsync(u => u.Id == request.userId);
+        user.PasswordHash = PasswordHashing.HashPassword(passwordHash, user.Salt, true);
+        await _lexBoxDbContext.SaveChangesAsync();
+        return Ok();
+    }
+}
